Move set_success and set_error to script/script_error.h
