+++ src/script/script_error.h
@@ -80,4 +80,21 @@ typedef enum ScriptError_t {
 
 const char *ScriptErrorString(const ScriptError error);
 
+namespace
+{
+inline bool set_success(ScriptError *ret)
+{
+    if (ret)
+        *ret = SCRIPT_ERR_OK;
+    return true;
+}
+
+inline bool set_error(ScriptError *ret, const ScriptError serror)
+{
+    if (ret)
+        *ret = serror;
+    return false;
+}
+
+} // anon namespace
 #endif // BITCOIN_SCRIPT_SCRIPT_ERROR_H
